+++ src/MILPSolver/MILPSolverCplex.h
@@ -83,7 +83,7 @@ class MILPSolverCplex: public IMILPSolver, public MILPSolverBase
 		}
 
 		virtual int increaseSolutionLimit(int increment);
-		virtual void setSolutionLimit(int limit);
+		virtual void setSolutionLimit(long limit);
 		virtual int getSolutionLimit();
 
 		virtual void setTimeLimit(double seconds);
