+++ src/MILPSolver/MILPSolverCplexLazy.h
@@ -37,7 +37,7 @@ class MILPSolverCplexLazy: public MILPSolverCplex
 		virtual E_ProblemSolutionStatus solveProblem();
 
 		virtual int increaseSolutionLimit(int increment);
-		virtual void setSolutionLimit(int limit);
+		virtual void setSolutionLimit(long limit);
 		virtual int getSolutionLimit();
 
 	private:
