+++ src/MILPSolver/MILPSolverOsiCbc.h
@@ -81,7 +81,7 @@ class MILPSolverOsiCbc: public IMILPSolver, MILPSolverBase
 		}
 
 		virtual int increaseSolutionLimit(int increment);
-		virtual void setSolutionLimit(int limit);
+		virtual void setSolutionLimit(long limit);
 		virtual int getSolutionLimit();
 
 		virtual void setTimeLimit(double seconds);
